requirements.txt
altair==6.0.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
backoff==2.2.1
bcrypt==5.0.0
blinker==1.9.0
build==1.4.0
cachetools==6.2.4
certifi==2026.1.4
charset-normalizer==3.4.4
chromadb==1.4.1
click==8.3.1
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.3.2
cycler==0.12.1
distro==1.9.0
durationpy==0.10
exceptiongroup==1.3.1
fastapi==0.128.0
filelock==3.20.0
flatbuffers==25.12.19
fonttools==4.61.1
fsspec==2025.12.0
gitdb==4.0.12
GitPython==3.1.46
google-ai-generativelanguage==0.6.15
google-api-core==2.29.0
google-api-python-client==2.188.0
google-auth==2.47.0
google-auth-httplib2==0.3.0
google-generativeai==0.8.6
googleapis-common-protos==1.72.0
groq==1.0.0
grpcio==1.76.0
grpcio-status==1.71.2
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httplib2==0.31.1
httptools==0.7.1
httpx==0.28.1
huggingface_hub==1.3.2
humanfriendly==10.0
idna==3.11
importlib_metadata==8.7.1
importlib_resources==6.5.2
Jinja2==3.1.6
joblib==1.5.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
kubernetes==35.0.0
lxml==6.0.2
markdown-it-py==4.0.0
MarkupSafe==2.1.5
matplotlib==3.10.8
mdurl==0.1.2
mmh3==5.2.0
mpmath==1.3.0
narwhals==2.15.0
networkx==3.4.2
numpy==2.2.6
oauthlib==3.3.1
onnxruntime==1.23.2
opentelemetry-api==1.39.1
opentelemetry-exporter-otlp-proto-common==1.39.1
opentelemetry-exporter-otlp-proto-grpc==1.39.1
opentelemetry-proto==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
orjson==3.11.5
overrides==7.7.0
packaging==24.2
pandas==2.3.3
pillow==12.0.0
pinecone==8.0.0
pinecone-plugin-assistant==3.0.1
pinecone-plugin-interface==0.0.7
posthog==5.4.0
proto-plus==1.27.0
protobuf==5.29.5
pyarrow==23.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybase64==1.4.3
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
Pygments==2.19.2
pyparsing==3.3.1
PyPika==0.50.0
pyproject_hooks==1.2.0
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.21
python-pptx==1.0.2
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
requests==2.32.5
requests-oauthlib==2.0.0
rich==14.2.0
rpds-py==0.30.0
rsa==4.9.1
scikit-learn==1.7.2
scipy==1.15.3
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
starlette==0.50.0
streamlit==1.53.0
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tokenizers==0.22.2
toml==0.10.2
tomli==2.4.0
torch==2.9.1+cpu
torchaudio==2.9.1+cpu
torchvision==0.24.1+cpu
tornado==6.5.4
tqdm==4.67.1
typer==0.21.1
typer-slim==0.21.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uritemplate==4.2.0
urllib3==2.6.3
uvicorn==0.40.0
watchdog==6.0.0
watchfiles==1.1.1
websocket-client==1.9.0
websockets==16.0
xgboost==3.1.3
xlsxwriter==3.2.9
zipp==3.23.0
